Adding friends

+++ db/migrate/20170809205313_create_restaurants.rb
@@ -2,8 +2,8 @@ class CreateRestaurants < ActiveRecord::Migration[5.1]
   def change
     create_table :restaurants do |t|
       t.string :placeId
-      t.decimal :lat
-      t.decimal :lng
+      t.float :lat
+      t.float :lng
       t.string :name
       t.string :address
       t.string :phone_number
